pyproject.toml
[tool.poetry]
name = "amazon-api"
version = "0.1.0"
description = ""
authors = ["Pamela Vermeer <vermeerp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-amazon-sp-api = "^1.9.10"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
